refactor: MVC 클래스 추상화
- MVC 클래스 추상화
- App 클래스 생성
