+++ src/js/core/Controller.js
@@ -10,21 +10,11 @@ export default class Controller {
     });
 
     this.bindEventHandlers();
-    console.log('controller loaded...');
-  }
-
-  bindEventHandlers() {
-    this.view.bindOnClickPaymentSubmit(this.purchase.bind(this));
-    this.view.bindOnClickNumberToggle();
-  }
-
-  purchase(amount) {
-    this.model.purchase(amount, (message) => {
-      this.updateView(message);
-    });
   }
 
   updateView(message) {
     this.view.update(message);
   }
+
+  bindEventHandlers() {}
 }
