+++ src/js/core/Model.js
@@ -1,0 +1,5 @@
+export default class Model {
+  init(callback) {
+    callback();
+  }
+}
