+++ src/js/index.js
@@ -1,10 +1,5 @@
-import Model from './Model/Model.js';
-import View from './View/View.js';
-import Controller from './Controller/Controller.js';
-import { $ } from './utils/utils.js';
+import App from './App.js';
 
-const model = new Model();
-const view = new View($('#app'));
-const controller = new Controller(model, view);
+const app = new App();
 
-controller.init();
+app.init();
